Simplify JS files
