+++ js/src/config.js
@@ -1,0 +1,6 @@
+window.Matte = {};
+Matte.Settings = {
+  "server_host":"127.0.0.1",
+  "server_port":8000,
+  "websockets_port":8001
+}
